remove configuration bin seperate methods to jwtUtils upgrade versions
